README.md added
Celebrate the spirit of Ramadan with our heartfelt project showcasing a beautiful "Ramadan Mubarak" greeting designed using HTML and CSS. Immerse yourself in the elegant Arabic typography and serene moon symbol, accompanied by a soulful Ramadan song, to spread joy and blessings to your Muslim followers. Whether you're looking to share festive greetings or explore the creative possibilities of web design, this project offers a touching tribute to the sacred month of Ramadan. Join us in embracing the spirit of unity, compassion, and devotion as we extend warm wishes of Ramadan Mubarak to all. Ramadan Kareem! 🌙✨